requirements.txt
fastapi
uvicorn
python-dotenv>=1.0.0
openai
requests>=2.31.0
python-docx
pdfkit
pyyaml>=6.0.1 
mcp[cli]
sseclient-py>=1.7.2
fastmcp>=1.13.0